+++ src/components/CommentsSection.js
@@ -1,10 +1,12 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { makeFriendly } from "../utils/num.js";
 
 import Comment from "./Comment.js";
+// stores
+import { useCommentsStore } from "./../stores/commentsStore.js";
 
 const CommentsSection = React.memo(
-	({ getComments, currentPostData, currentComments, setCurrentComments }) => {
+	({ currentPostData, currentComments, setCurrentComments, permalink }) => {
 		const [comments, setComments] = useState(currentComments);
 
 		useEffect(() => {
@@ -24,6 +26,8 @@ const CommentsSection = React.memo(
 						setComments={setComments}
 						comments={comments}
 						currentComments={currentComments}
+						permalink={permalink}
+						key="sortByButton"
 					></SortByButton>
 				</span>
 				{/*todo: show loading comments banner */}
@@ -32,7 +36,6 @@ const CommentsSection = React.memo(
 						if (commentObj.kind === "more") return null;
 						return (
 							<Comment
-								getComments={getComments}
 								data={commentObj.data}
 								topLevel={true}
 								key={commentObj.data.id}
@@ -48,15 +51,23 @@ const CommentsSection = React.memo(
 	}
 );
 
-const SortByButton = ({ setComments, comments, currentComments }) => {
+const SortByButton = ({
+	setComments,
+	comments,
+	currentComments,
+	permalink
+}) => {
 	/*
 	todo
-	sort by datalist,
-	1/3 * 98vw into a variable to use for photos/videos ???
+	remember sortBy for different posts
+	selected value getshighlighted but comments dont get sorted accordingly.
+
+	same thing need to be done for collaping comments.
 	*/
+	const { sortBy, setSortBy, changeCommentObj } = useCommentsStore();
 
-	const [sortBy, setSortBy] = useState(null);
-
+	// ?sort=new
+	// https://www.reddit.com/r/pics/comments/nqiod2/the_columbia_crew_in_orbit_recovered_from/.json?sort=new
 	const style = {
 		position: "relative",
 		top: "5px",
@@ -71,78 +82,87 @@ const SortByButton = ({ setComments, comments, currentComments }) => {
 
 	// not working currently
 	const sortingMethods = {
+		// ya auto doesnt work anymore now.
 		auto: () => {
 			setComments(currentComments);
+			return currentComments;
 		},
 		upvotes: () => {
+			let l = [];
 			setComments((coms) => {
 				let c = [...coms];
-				return c.sort(({ data: data1 }, { data: data2 }) => {
+				l = c.sort(({ data: data1 }, { data: data2 }) => {
 					return data2.ups - data1.ups;
 				});
+				changeCommentObj(permalink, { link: permalink, comments: l });
+				return l;
 			});
+			return l;
 		},
 		new: () => {
+			let l = [];
 			setComments((coms) => {
 				let c = [...coms];
-				return c.sort(({ data: data1 }, { data: data2 }) => {
+				l = c.sort(({ data: data1 }, { data: data2 }) => {
 					return data2.created_utc - data1.created_utc;
 				});
+				changeCommentObj(permalink, { link: permalink, comments: l });
+				return l;
 			});
+			return l;
 		},
 		old: () => {
 			console.log(comments);
+			let l = [];
 			setComments((coms) => {
-				// return [];
 				let c = [...coms];
-				return c.sort(({ data: data1 }, { data: data2 }) => {
+				l = c.sort(({ data: data1 }, { data: data2 }) => {
 					return data1.created_utc - data2.created_utc;
 				});
+				changeCommentObj(permalink, { link: permalink, comments: l });
+				return l;
 			});
+			return l;
 		},
 		awards: () => {
 			console.log(comments);
+			let l = [];
 			setComments((coms) => {
-				// we need immer for this now !!
-				// also read wesbos's array destructuring articles
-				// let c = [...coms];
-				// c.shift();
-				// return c;
 				let c = [...coms];
-				return c.sort(({ data: data1 }, { data: data2 }) => {
-					console.log(data1, data2);
-					console.log(
-						data1.total_awards_received -
-							data2.total_awards_received
-					);
-					return (
-						data1.total_awards_received -
-						data2.total_awards_received
-					);
-				});
+				l = c.sort(
+					({ data: data1 }, { data: data2 }) =>
+						data2.total_awards_received -
+						data1.total_awards_received
+				);
+				changeCommentObj(permalink, { link: permalink, comments: l });
+				return l;
 				// console.log(sorted);
 				// return sorted;
 			});
+			// l is [] because of delayed assignment.
+			return l;
 		}
 	};
 
 	const sort = (e) => {
-		sortingMethods[e.target.value]();
+		let comms = sortingMethods[e.target.value]();
+		// console.log("TESTSYT");
+		// console.log(comms);
+		// changeCommentObj(permalink, { link: permalink, comments: comms });
+		setSortBy({
+			[permalink]: e.target.value
+		});
 	};
 
 	return (
 		<select name="sortby" id="sortby" style={style} onChange={sort}>
 			{Object.keys(sortingMethods).map((name) => (
-				<option value={name}>{name}</option>
+				<option value={name} selected={sortBy[permalink] === name}>
+					{name}
+				</option>
 			))}
 		</select>
 	);
 };
 
 export default CommentsSection;
-
-/*
-SORTING NOT WORKING
-
-since when do the award icons in comments go so high ??
-*/