+++ src/components/FocusView.js
@@ -1,29 +1,30 @@
 import { useHotkeys } from "react-hotkeys-hook";
-import { useEffect, useState, useCallback, useRef } from "react";
-import {
-	useHistory,
-	useLocation,
-	useParams,
-	useRouteMatch
-} from "react-router-dom";
+import { useEffect, useState } from "react";
+import { useHistory, useRouteMatch } from "react-router-dom";
+
+// components
 import Post from "./Post.js";
+import CommentsSection from "./CommentsSection.js";
 
 // stores
 import { useViewStyleStore } from "./../stores/viewStyle.js";
-import CommentsSection from "./CommentsSection.js";
+import { useCommentsStore } from "./../stores/commentsStore.js";
 
 const FocusView = ({
 	postsData,
 	getComments,
-	initPostNo = 0,
+	initPostNo = { current: 0 },
 	viewStyle,
 	shouldBlurAll,
 	postsSeen,
 	setPostsSeen,
-	setLastSeen
+	setLastSeen,
+	expandView,
+	loadListings,
+	loaded
 }) => {
 	const history = useHistory();
-	const [currentPost, setCurrentPost] = useState(initPostNo);
+	const [currentPost, setCurrentPost] = useState(initPostNo.current);
 	const [currentComments, setCurrentComments] = useState([]); // {url:[url], comments: obj `children[]]`}
 	const currentPostData = postsData[currentPost]?.data;
 	const [shouldBlur, setBlur] = useState(
@@ -53,15 +54,15 @@ const FocusView = ({
 		params: { subreddit }
 	} = useRouteMatch("/r/:subreddit");
 
+	// encode here or decode here ??
 	let permalink = currentPostData?.permalink
-		? `https://www.reddit.com${currentPostData.permalink}`
+		? `https://www.reddit.com${currentPostData?.permalink}`
 		: window.location.pathname.replace(`/r/${subreddit}/`, "");
 
-	// const location = useLocation();
-	// console.log({ location });
-
-	// const { permalink } = useParams();
-	console.log({ permalink });
+	// comments changing on resort logic
+	const sortBy = useCommentsStore((s) => s.sortBy);
+	// ! clear store data on sub change ?
+	// ! omg check comments ... ....
 	// https://github.com/jaywcjlove/hotkeys/#defining-shortcuts
 	useHotkeys(
 		"n",
@@ -77,6 +78,7 @@ const FocusView = ({
 					history.replace(
 						`/r/${subreddit}/https://www.reddit.com${postsData[c].data.permalink}`
 					);
+					expandView(c);
 					return c;
 				}
 			});
@@ -105,6 +107,7 @@ const FocusView = ({
 					history.replace(
 						`/r/${subreddit}/https://www.reddit.com${postsData[c].data.permalink}`
 					);
+					expandView(c);
 					return c;
 				}
 			});
@@ -112,29 +115,44 @@ const FocusView = ({
 		// practically we should load more posts at this point. or show a msg when the listing has been finished
 	);
 
+	// use useEffect caches the values not in the dep array ? hmmm;
 	useEffect(() => {
 		// realistically this should happen when the sub changes.
-		return () => setUnBlurred({});
+		console.warn("FOCUSVIEW mounted!");
+		return () => {
+			console.log("FOCUSVIEW unmounted !");
+			setUnBlurred({});
+			// ohmygod wow this fixed the issue;
+			/* basically post would not load when going from focusView to StackView (next sub) -> 
+			browsing in FocusView -> hitting back 2x to get back to the first FocusView post.*/
+			initPostNo.current = 0;
+		};
 	}, []);
+
 	useEffect(() => {
 		// Load Comments
 		// let curl = `${"https://www.reddit.com/r/Showerthoughts/comments/mw2amn/having_to_attend_a_wedding_you_dont_want_to_sucks/"}.json`;
 		// todo: change url to permalink in above line
 		let controller = new AbortController();
-		const result = getComments(permalink, controller.signal);
-		result.then(({ comObj, data, aborted }) => {
-			if (aborted) return null;
-			// todo we need better error handling lol.
-			if (comObj === 1) alert("likely fetch request went wrong");
-			console.log(comObj);
-			setCurrentComments(comObj.comments);
-			// for singel page loads
-			if (shouldBlur === undefined) {
-				setBlur(shouldBlurAll && data.over_18);
-			}
-		});
+		setTimeout(() => {
+			const result = getComments(permalink, controller.signal);
+			result.then(({ comObj, data, aborted }) => {
+				if (aborted) return null;
+				// todo we need better error handling lol.
+				if (comObj === 1) alert("likely fetch request went wrong");
+				if (comObj.comments) setCurrentComments(comObj.comments);
+				else debugger;
+				console.log({ comObj });
+				// for singel page loads
+				if (shouldBlur === undefined) {
+					setBlur(shouldBlurAll && data?.over_18);
+				}
+			});
+		}, 1);
 		return () => {
 			// abort here yes !
+			// ok nice !! do this in another effect with only currentPost as dependency.
+			// & store controller in ref ?
 			controller.abort();
 		};
 		// I HAVE GONE THROUGH A LOT OF TROUBLE FOR THIS UNREAL PROBLEM
@@ -142,7 +160,11 @@ const FocusView = ({
 		// BUT THEN HOW DO YOU KNOW IF YOY SHOULD USE LOCALSTATE OR REQUEST NEW ???
 		// ? WELL NOW I GUESS ITS A BAD IDEA, JUST USE ANOTHER USESTATE FOR THE SUBREDDIT NAME.
 		// ? AND SET IT WHEN THIS UNMOUNTS.
-	}, [currentPost]);
+	}, [postsData, currentPost, sortBy[currentPostData?.permalink]]);
+	// }, [currentPost, sortBy[currentPostData?.permalink]]);
+	// we are on the right track
+	// }, [postsData, currentPost, sortBy[currentPostData?.permalink]]);
+	// added currentPostData?.permalink above.
 
 	if (currentPostData === undefined) {
 		return <h1 style={{ textAlign: "center" }}>Loading Post</h1>;
@@ -165,11 +187,11 @@ const FocusView = ({
 			/>
 			<CommentsSection
 				{...{
-					getComments,
 					currentPostData,
 					currentComments,
 					setCurrentComments
 				}}
+				permalink={currentPostData.permalink}
 			></CommentsSection>
 		</>
 	);