+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## CHANGELOG
 
-1. images and videos display fixed
+1. fixed sorting comments by awards.
+2. fixed the not working lastSeenPost thing with n and p.
+3. fixed the bad history.goBack issue.
+4. fixed the postsData not loading when going back to StackView from FocusView !!
+5. fixed the bad history and fixed a lot of issues, code is pretty bad rn.
+
+## bug
+
+whenever we try to sort comments the no of comment objects increase to thousands eventually.